pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wp-translation"
version = "1.0.0"
description = "Fine-tune LLMs for WordPress translation using data from translate.wordpress.org"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "WordPress Translation Team"}
]
keywords = ["wordpress", "translation", "llm", "fine-tuning", "qlora", "mistral"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "torch>=2.2.0",
    "transformers>=4.38.0",
    "datasets>=2.17.0",
    "accelerate>=0.27.0",
    "peft>=0.8.2",
    "trl>=0.7.10",
    "bitsandbytes>=0.42.0",
    "unbabel-comet>=2.2.0",
    "sacrebleu>=2.4.0",
    "polib>=1.2.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.0",
    "tenacity>=8.2.0",
    "typer[all]>=0.9.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.4.0",
    "tensorboard>=2.15.0",
    "tqdm>=4.66.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.scripts]
wp-translate = "cli.main:app"

[project.urls]
Homepage = "https://github.com/wordpress/wp-translation-llm"
Documentation = "https://github.com/wordpress/wp-translation-llm#readme"
Repository = "https://github.com/wordpress/wp-translation-llm"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "cli*"]

[tool.setuptools.package-dir]
"" = "."

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
